+++ test/unit/resources/config_property_spec.rb
@@ -39,5 +39,20 @@ describe 'ConfigProperty' do
       config_property = RubyAem::Resources::ConfigProperty.new(@mock_client, 'alias', 'String', '/crx/server')
       config_property.create('org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet')
     end
+
+    it 'should call client with expected parameters when property is a keyword for org.apache.http.proxyconfigurator.config' do
+      expect(@mock_client).to receive(:call).once.with(
+        RubyAem::Resources::ConfigProperty,
+        'createapachehttpcomponentsproxyconfiguration',
+        name: 'proxy_host',
+        config_node_name: 'org.apache.http.proxyconfigurator.config',
+        type: 'String',
+        value: '192.168.1.1',
+        proxy_host: '192.168.1.1',
+        proxy_host_type_hint: 'String'
+      )
+      config_property = RubyAem::Resources::ConfigProperty.new(@mock_client, 'proxy_host', 'String', '192.168.1.1')
+      config_property.create('org.apache.http.proxyconfigurator.config')
+    end
   end
 end